+++ src/job/infrastructure/repositories/job-repository.ts
@@ -5,7 +5,7 @@ import { IEventStore } from '@core/IEventStore';
 import { Job } from '@domain/job/job';
 import { IJobRepository } from '@domain/job/job-repository.interface';
 
-import { Repository } from './repository';
+import { Repository } from '../../../infrastructure/repositories/repository';
 
 @injectable()
 export class JobRepository extends Repository<Job> implements IJobRepository {
